+++ spec/presenters/offer_presenter_spec.rb
@@ -42,4 +42,36 @@ describe OfferPresenter do
     end
   end
 
+  describe "#description" do
+    subject { presenter.description }
+
+    before do
+      offer.card = Card.new(currency: Currency.new(name: "Dinero"))
+      offer.points_awarded = 7_500
+    end
+
+    context "when 'condition' is 'on first purchase'" do
+      it "prints reasonable description" do
+        offer.condition = "on_first_purchase"
+        is_expected.to eq "7,500 Dinero points awarded upon making your first purchase using this card."
+      end
+    end
+
+    context "when 'condition' is 'on approval'" do
+      it "prints reasonable description" do
+        offer.condition = "on_approval"
+        is_expected.to eq "7,500 Dinero points awarded upon a successful application for this card."
+      end
+    end
+
+    context "when 'condition' is 'on minimum spend'" do
+      it "prints reasonable description" do
+        offer.condition = "on_minimum_spend"
+        offer.spend = 4_500
+        offer.days  = 40
+        is_expected.to eq "Spend $4,500.00 within 40 days to receive a bonus of 7,500 Dinero points"
+      end
+    end
+  end
+
 end